updated wrapper

+++ wrapper.py
@@ -45,7 +45,7 @@ def toCDoubleArray(arr):
     return (ctypes.c_double * len(arr))(*arr)
 
 def CDoubletoPyArray(arr, length):
-    return ctypes.cast(length, ctypes.POINTER(ctypes.c_int * length)).contents # need to test
+    return ctypes.cast(length, ctypes.POINTER(ctypes.c_double * length)).contents() # need to test
 
 def getNetworkStruct(args):
     pass
